changes from comments
