+++ care/emr/resources/role/spec.py
@@ -1,7 +1,8 @@
-from pydantic import UUID4
+from pydantic import UUID4, BaseModel, ValidationError, model_validator
 
 from care.emr.resources.base import EMRResource
 from care.security.models import PermissionModel, RoleModel
+from care.security.permissions.base import PermissionController
 
 
 class PermissionSpec(EMRResource):
@@ -38,3 +39,32 @@ class RoleReadSpec(RoleBaseSpec):
         mapping["id"] = obj.external_id
         mapping["permissions"] = obj.get_permissions_for_role()
         return mapping
+
+
+class PermissionManageSpec(BaseModel):
+    permissions: list[str]
+
+    @model_validator(mode="after")
+    def validate_permissions(self):
+        valid_permissions = PermissionController.get_permissions().keys()
+        self.permissions = list(set(self.permissions))  # Remove duplicates
+        for permission in self.permissions:
+            if permission not in valid_permissions:
+                error = f"Invalid permission slug: {permission}."
+                raise ValidationError(error)
+        return self
+
+
+class RoleConfig(PermissionManageSpec):
+    role: RoleCreateSpec
+
+    @model_validator(mode="after")
+    def validate_role_and_permissions(self):
+        system_roles = RoleModel.objects.filter(is_system=True).values_list(
+            "name", flat=True
+        )
+        role = self.role.name
+        if role in system_roles:
+            error = f"Role {role} already exists."
+            raise ValidationError(error)
+        return self
